LayoutTests/http/tests/site-isolation/fullscreen.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script>
    if (window.testRunner) {
        testRunner.waitUntilDone();
        testRunner.dumpAsText()
        testRunner.dumpFullScreenCallbacks();
    }
    addEventListener("message", (event) => {
        document.getElementById("mylog").innerHTML += event.data + "<br>";
        document.getElementById("mylog").innerHTML += 'iframe border style after transition: ' + getComputedStyle(document.getElementById("frame")).getPropertyValue('border') + "<br><br>";
        if (event.data.startsWith('Size after exiting') && window.testRunner) { testRunner.notifyDone() }
    });
</script>
<iframe id='frame' allowfullscreen src="http://localhost:8000/site-isolation/resources/fullscreen.html" frameborder=0></iframe>
<div id=mylog>
FIXME: Size after entering should be 600x800 like it is with site isolation off.<br>
The size currently comes from screenRectOfContents.<br>
Also, the iframe's border 'inset' style should be 'none' after entering fullscreen.<br><br>
</div>
