LayoutTests/http/tests/site-isolation/fullscreen-expected.txt

supportsFullScreen() == true
enterFullScreenForElement()
beganEnterFullScreen() - initialRect.size: {300, 150}, finalRect.size: {300, 150}
exitFullScreenForElement()
beganExitFullScreen() - initialRect.size: {300, 150}, finalRect.size: {300, 150}

FIXME: Size after entering should be 600x800 like it is with site isolation off.
The size currently comes from screenRectOfContents.
Also, the iframe's border 'inset' style should be 'none' after entering fullscreen.

Size after entering fullscreen: 150x300
iframe border style after transition: 0px inset rgb(0, 0, 0)

Size after exiting fullscreen: 150x300
iframe border style after transition: 0px inset rgb(0, 0, 0)
